webhook_requirements.txt
Flask==2.3.3
pycups==2.0.1
requests==2.31.0
